% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm.R
\docType{package}
\name{fcm}
\alias{fcm}
\alias{fcm-package}
\title{Inference of Fuzzy Cognitive Maps (FCMs)}
\description{
Estimates the Inference of a Fuzzy Cognitive Map. Provides a selection of 3 different inference rules (including 3 more, the clamped types of inference rules) and 4 threshold functions in order to obtain the FCM inference procedure. FCM (Kosko, 1986) is proven to be capable of causal inference and is applicable to complex decision problems where numerous interlinked dependent variables influence one another. The fcm package includes the fcm.infer function which returns a dataframe of the concepts' values of each state after the inference procedure.
}
